+++ samples/JS/decorators/SObject.js
@@ -1,0 +1,20 @@
+import {
+    Model,
+    Column,
+    Table,
+    CreatedAt,
+    UpdatedAt,
+} from "sequelize-typescript"
+
+const some = null;
+@Table({
+    tableName: "SObjects",
+    any: some,
+})
+export class {
+    @CreatedAt
+    @Column
+    createdAt() {
+        return null
+    }
+}