+++ samples/JS/decorators/decorators.ts
@@ -4,8 +4,7 @@ import { suite, test } from '@testdeck/mocha';
 class Suite {
 
   @test('a custom name for the test')
-  test() {
-  }
+  test() {}
 
   @test 'alternate naming'() { }
 }